% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciplot.gg.R
\name{ciplot.gg}
\alias{ciplot.gg}
\title{Draw 95\% Confidence Interval}
\usage{
ciplot.gg(lfpcrobj, d)
}
\arguments{
\item{lfpcrobj}{lfpcrobj}

\item{d}{Bootstrapped coefficients}
}
\description{
Draw 95\% Confidence Interval
}
\examples{
re<-hd_lfpca(Ydat,T,J,I,visit, varthresh=0.85, timeadjust=FALSE)
lpcr_lasso(Y,lfpca=re)
}
\author{
Liwen Wu, Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\references{
TBA
}
\keyword{glmnet}
\keyword{hdlfpca}

