% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_lfpca.R
\name{hd_lfpca}
\alias{hd_lfpca}
\title{Conduct longitudinal principal component analysis}
\usage{
hd_lfpca(Y, T, J, I, visit, verbose = 1, prefix = date(), Nx = NA,
  Nw = NA, varthresh = 0.99, projectthresh = 1, timeadjust = FALSE,
  figure = FALSE)
}
\arguments{
\item{Y}{A number}

\item{T}{A number}

\item{J}{A number}

\item{I}{A number}

\item{visit}{A number}

\item{verbose}{A number}

\item{Nx}{A number}

\item{Nw}{A number}

\item{varthresh}{A number}

\item{projectthresh}{A number}

\item{timeadjust}{A number}
}
\value{
xi

phix0

phix1

zeta

phiw
}
\description{
Conduct longitudinal principal component analysis
}
\examples{
re<-hd_lfpca(Y,T,J,I,visit, varthresh=0.85, timeadjust=FALSE)
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\references{
Zipunnikov, V., Greven, S., Shou, H., Caffo, B., Reich, D. S., & Crainiceanu, C. (2014). Longitudinal High-Dimensional Principal Components Analysis with Application to Diffusion Tensor Imaging of Multiple Sclerosis. The Annals of Applied Statistics, 8(4), 2175–2202.
}
\keyword{glmnet}
\keyword{hdlfpca}

