% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfpcr_boot.R
\name{lfpcr_boot}
\alias{lfpcr_boot}
\title{Run boot strapping to compute 95\% CI}
\usage{
lfpcr_boot(datct = NULL, idlist = NULL, B = 100, idvar = NULL,
  timevar = NULL, inYrs = FALSE, predvarlist = NULL, outcomevar = NULL,
  covariates = NULL, varthresh = 0.85, lambdalist = 2^(c(-8:8)),
  penalty.factor = NULL, nfold = 10, multicore = FALSE, mc.cores = 2)
}
\arguments{
\item{idlist}{List of sample unit}

\item{idvar}{ID variable name from the dataset}

\item{timevar}{time variable name from the dataset in days from the baseline}

\item{inYrs}{whether the time variable needs to be converted in years, default FALSE}

\item{predvarlist}{columns of the predictors}

\item{outcomevar}{outcome variable}

\item{varthresh}{threshold for LFPCA}

\item{multicore}{whether conducting multicore computation (default=FALSE)}

\item{mc.cores}{number of cores (default=2)}

\item{seednum}{random seed number}
}
\description{
Run boot strapping to compute 95\% CI
}
\author{
Seonjoo lee \email{sl2670@cumc.columbia.edu}
}
\references{
TBA
}

