% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfpcr_boot_oneunit.R
\name{lfpcr_boot_oneunit}
\alias{lfpcr_boot_oneunit}
\title{Run one unit boot strapping}
\usage{
lfpcr_boot_oneunit(datct = NULL, idlist = NULL, seednum = 1000,
  idvar = NULL, timevar = NULL, inYrs = FALSE, predvarlist = NULL,
  outcomevar = NULL, covariates = NULL, varthresh = 0.85,
  lambdalist = 2^(c(-8:8)), penalty.factor = NULL, nfold = 10)
}
\arguments{
\item{idlist}{List of sample unit}

\item{seednum}{random seed number}

\item{idvar}{ID variable name from the dataset}

\item{timevar}{time variable name from the dataset in days from the baseline}

\item{inYrs}{whether the time variable needs to be converted in years, default FALSE}

\item{predvarlist}{columns of the predictors}

\item{outcomevar}{outcome variable}

\item{varthresh}{threshold for LFPCA}
}
\description{
Run one unit boot strapping
}
\references{
TBA
}
\author{
Seonjoo lee \email{sl2670@cumc.columbia.edu}
}
