% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcr.backward.R
\name{lpcr.backward}
\alias{lpcr.backward}
\title{Conduct longitudinal principal component regression with model selection}
\usage{
lpcr.backward(Y, Xmat = NULL, lfpca = NULL, cov = NULL, T = NULL,
  J = NULL, I = NULL, visit = NULL, varthresh = 0.85,
  timeadjust = FALSE, nfold = 10, lambdalist = 2^(c(-10:10)/2),
  penalty.factor = NULL, M = NULL, method = "AUC", seednum = 1234)
}
\arguments{
\item{Y}{Outcome in PCR}

\item{Xmat}{Longitudinal Predictor Matrix}

\item{lfpca}{(default=NULL) LFPC results. If it is not specified, the following parameters have to be specified to run LPCA}

\item{cov}{Covariate}

\item{T}{Time of theimage collection}

\item{J}{Total number of observations}

\item{I}{Total number of subjects}

\item{visit}{Vector of number of visits per subjects}

\item{varthresh}{(default=0.99) Threshold for variance explained for both subject-specific and subject-visit specific compoents for dimension selection}

\item{timeadjust}{(default=TRUE) Scale time per subject}

\item{method}{(default=AUC) backward selection criteria. c('AUC','bic','aic')}

\item{verbose}{(default=FALSE)}

\item{Nx}{Dimension of the subject-specific components}

\item{Nw}{Dimension of the subject-visit specific components}

\item{projectthresh}{Threshold for variance explain in the first step of SVD}
}
\value{
xi

phix0

phix1

zeta

phiw
}
\description{
Conduct longitudinal principal component regression with model selection
}
\examples{
I=200
visit=rpois(I,1)+3
time = unlist(lapply(visit, function(x) scale(c(0,cumsum(rpois(x-1,1)+1)))))
J = sum(visit)
V=1000
phix0 = matrix(0,V,3);phix0[1:50,1]<-.1;phix0[1:50 + 50,2]<-.1;phix0[1:50 + 100,3]<-.1
phix1 = matrix(0,V,3);phix1[1:50+150,1]<-.1;phix1[1:50 + 200,2]<-.1;phix1[1:50 + 250,3]<-.1
phiw = matrix(0,V,3);phiw[1:50+300,1]<-.1;phiw[1:50 + 350,2]<-.1;phiw[1:50 + 400,3]<-.1
xi = t(matrix(rnorm(I*3),ncol=I)*c(8,4,2))*3
zeta = t(matrix(rnorm(J*3),ncol=J)*c(8,4,2))*2
Xmat = phix0\%*\% t(xi[rep(1:I, visit),]) + phix1\%*\% t(time * xi[rep(1:I, visit),]) + phiw \%*\% t(zeta) + matrix(rnorm(V*J,0,.1),V,J)
beta=c(1,-1,1)/10
p=exp( - xi \%*\% beta)/(1 + exp( - xi \%*\% beta))
Y = unlist(lapply(p,function(x)rbinom(1,1,prob=x)))
re<-hd_lfpca(Ydat,Xmat,T,J,I,visit, varthresh=0.85, timeadjust=FALSE)
lpcr.backward(Y,lfpca=re)
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{hdlfpca}
