% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcr_lasso.R
\name{lpcr_lasso}
\alias{lpcr_lasso}
\title{Conduct longitudinal principal component regression analysis}
\usage{
lpcr_lasso(Y, T = NULL, J = NULL, I = NULL, visit = NULL,
  lfpca = NULL, cov = NULL, nfold = 10, lambdalist = 2^(c(-10:10)/2),
  penalty.factor = NULL, M = NULL, seednum = 1234)
}
\arguments{
\item{Y}{A number}

\item{T}{A number}

\item{J}{A number}

\item{I}{A number}

\item{visit}{A number}

\item{cov}{Covariate}

\item{varthresh}{A number}

\item{projectthresh}{A number}

\item{timeadjust}{A number}
}
\value{
xi

phix0

phix1

zeta

phiw
}
\description{
Conduct longitudinal principal component regression analysis
}
\examples{
re<-hd_lfpca(Ydat,T,J,I,visit, varthresh=0.85, timeadjust=FALSE)
lpcr_lasso(Y,lfpca=re)
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{hdlfpca}

