% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciplot.gg.R
\name{ciplot.gg}
\alias{ciplot.gg}
\title{Draw 95\% Confidence Interval}
\usage{
ciplot.gg(lfpcrobj, d)
}
\arguments{
\item{lfpcrobj}{lfpcrobj}

\item{d}{Bootstrapped coefficients}
}
\description{
Draw 95\% Confidence Interval
}
\author{
Liwen Wu, Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\references{
TBA
}
\keyword{glmnet}
\keyword{hdlfpca}

