% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciplot.gg.R
\name{ciplot.gg}
\alias{ciplot.gg}
\title{Draw 95\% Confidence Interval}
\usage{
ciplot.gg(lfpcrobj, d, exp = FALSE, yrange = c(-5, 5))
}
\arguments{
\item{lfpcrobj}{lfpcrobj}

\item{d}{Bootstrapped coefficients}

\item{exp}{Whether take exponential for the estimated CIs.}

\item{yrange}{set up range for the 95% CIs.}
}
\description{
Draw 95\% Confidence Interval
}
\references{
TBA
}
\author{
Liwen Wu, Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{hdlfpca}
